Code/Website-LandingPage/requirements.txt
flask==3.1.3
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0